let us create a base serverconfig to get the settings
