+++ solid/lib/ServerConfig.php
@@ -4,11 +4,12 @@
 	use OCP\IConfig;
 	use OCP\IUserManager;
 	use OCP\IUrlGenerator;
-
+	use OCA\Solid\BaseServerConfig;
+	
 	/**
 	 * @package OCA\Solid
 	 */
-	class ServerConfig {
+	class ServerConfig extends BaseServerConfig {
 		private IConfig $config;
 		private IUrlGenerator $urlGenerator;
 		private IUserManager $userManager;
@@ -22,41 +23,6 @@
 			$this->config = $config;
 			$this->userManager = $userManager;
 			$this->urlGenerator = $urlGenerator;
-		}
-
-		/**
-		 * @return string
-		 */
-		public function getPrivateKey() {
-			$result = $this->config->getAppValue('solid','privateKey');
-			if (!$result) {
-				// generate and save a new set if we don't have a private key;
-				$keys = $this->generateKeySet();
-				$this->config->setAppValue('solid','privateKey',$keys['privateKey']);
-				$this->config->setAppValue('solid','encryptionKey',$keys['encryptionKey']);
-			}
-			return $this->config->getAppValue('solid','privateKey');
-		}
-
-		/**
-		 * @param string $privateKey
-		 */
-		public function setPrivateKey($privateKey) {
-			$this->config->setAppValue('solid','privateKey',$privateKey);
-		}
-
-		/**
-		 * @return string
-		 */
-		public function getEncryptionKey() {
-			return $this->config->getAppValue('solid','encryptionKey');
-		}
-
-		/**
-		 * @param string $publicKey
-		 */
-		public function setEncryptionKey($publicKey) {
-			$this->config->setAppValue('solid','encryptionKey',$publicKey);
 		}
 
 		/**
@@ -196,23 +162,5 @@
 				$user = $this->userManager->get($userId);
 				$user->setDisplayName($fields['name']);
 			}
-		}
-		private function generateKeySet() {
-			$config = array(
-				"digest_alg" => "sha256",
-				"private_key_bits" => 2048,
-				"private_key_type" => OPENSSL_KEYTYPE_RSA,
-			);
-			// Create the private and public key
-			$key = openssl_pkey_new($config);
-
-			// Extract the private key from $key to $privateKey
-			openssl_pkey_export($key, $privateKey);
-			$encryptionKey = base64_encode(random_bytes(32));
-			$result = array(
-				"privateKey" => $privateKey,
-				"encryptionKey" => $encryptionKey
-			);
-			return $result;
 		}
 	}
